+++ tests/PokeGame.IntegrationTests/Evolutions/EvolutionIntegrationTests.cs
@@ -10,7 +10,6 @@ using PokeGame.Core.Forms;
 using PokeGame.Core.Items;
 using PokeGame.Core.Moves;
 using PokeGame.Core.Permissions;
-using PokeGame.Core.Pokemon;
 using PokeGame.Core.Regions;
 using PokeGame.Core.Search;
 using PokeGame.Core.Species;
@@ -265,7 +264,7 @@ public class EvolutionIntegrationTests : IntegrationTests
   {
     UpdateEvolutionPayload payload = new()
     {
-      Level = new Optional<int?>(0)
+      Level = new Optional<byte?>(0)
     };
 
     await Assert.ThrowsAsync<ValidationException>(async () => await _evolutionService.UpdateAsync(_evolution.EntityId, payload));
@@ -298,7 +297,7 @@ public class EvolutionIntegrationTests : IntegrationTests
   {
     UpdateEvolutionPayload payload = new()
     {
-      Level = new Optional<int?>(20),
+      Level = new Optional<byte?>(20),
       Friendship = false,
       Gender = new Optional<Gender?>(Gender.Male),
       ItemId = new Optional<Guid?>(_item.EntityId),
@@ -310,7 +309,7 @@ public class EvolutionIntegrationTests : IntegrationTests
     Assert.NotNull(evolution);
     Assert.Equal(_evolution.EntityId, evolution.Id);
     Assert.Equal(3, evolution.Version);
-    Assert.Equal(20, evolution.Level);
+    Assert.Equal((byte)20, evolution.Level);
     Assert.False(evolution.Friendship);
     Assert.Equal(Gender.Male, evolution.Gender);
     Assert.NotNull(evolution.Item);
